requirements.txt
termcolor==2.3.0
pyOpenSSL==23.1.1